Set current unix timestamp as version number
If the version number is not explicitly set, BOSH simply increases this each time. This means, if the current final version number for the name foobar-sha1 is 5 and the current dev version number is 3 (foobar-sha1-5+dev.3.yml), the next dev version number will be 4 (foobar-sha1-5+dev.4.yml).
The problem here is that the dev releases usually are not in the repository, so the dev version number created by BOSH in the Concourse container will always be 1 (foobar-sha1-5+dev.1.yml). So just to make sure we use the current timestamp instead of automatically increasing the last dev version number, which makes no sense in Concourse.