Return local job properties and global properties
The newest BOSH directors support BOSH v2 manifests only and will not
deploy a manifest containing global properties anymore.
Therefore, we collect all local job properties and the global ones and
merge them together in one hash for the time being.
In the future we'll need to get rid of `properties` or at least access
properties similar to the Ops file path structure.
The unit tests were capable of handling local and global property
manifests separately. The added unit test allows for manifests that
contain both global and local properties. This is needed until the
migration to BOSH v2 is complete.

Co-authored-by: Robert Gogolok <gogolok@gmail.com>