Added find_in_deployment_manifest method
The find_in_deploymnet_manifest method fetches the manifest of
a already deployed deployment and return a part of that manifest.
The returned part is defined by an ops-like search string.
The default value of that search string is nil.
If the search string is nil then it returns the whole manifest.