Remove default arguments on BOSH stop
We are removing the --skip-drain from default because some of our BOSH
releases depend on skip drain on some scenarios where otherwise a
process would timeout while stopping.

We should also consider that with this flag, we do not test the drain
script developed on our BOSH releases.

The --hard is being removed to speed up the unit-tests execution and to
have a more natural execution of the tests, as usually the operator does
not stop its instances with this flag on.