[Maintenance] Description of the stop_instance Method
We've had the `--hard` flag as a default parameter for this method. It turned out that using the `--hard`
flag did not result in the expected behaviour. It did create a BOSH task which succeeded, but it did
not stop the VM.

We're now removed the `--hard` and `--skip-drain` as default parameters that, when using the `stop_instance`
method, the VM is actually stopped.