Supports '?' to indicate optional key
When looking for a path that is not in the manifest, find will throw an
exception by default.

This commit introduces support to '?' to indicate that a property might
not exist, which in this case it should not throw an exception, but
return nil.

This is particularly useful when testing one test case against different
manifests (e.g. a9s PostgreSQL unit-tests).